// nextjs/gbiz-next/src/packages/request/types.ts
export enum RuntimeEnvironment {
  /** 浏览器环境 */
  BROWSER = 1,
  /** SSR服务端环境 */
  SSR = 2,
  APP = 3,
  WEBWORKER = 4,
  RN = 5,
}

interface Storage {
  getItem(key: string): string | null
  setItem(key: string, value: string): void
}

export interface FingerPrintUrl {
  url: string
  event?: string
  method?: string
}

export interface CreateHttpClientConfig extends CreateAxiosDefaults {
  /** 额外的请求拦截器 */
  requestInterceptors?: RequestInterceptor[]
  /** 额外的响应拦截器 */
  responseInterceptors?: ResponseInterceptor[]
  /** 禁用全部内置的拦截器，默认false */
  disablePresetInterceptors?: boolean
  /** 禁用指定名称的内置请求拦截器 */
  disableRequestPresetInterceptors?: string[]
  /** 禁用指定名称的内置响应拦截器 */
  disableResponsePresetInterceptors?: string[]
  /** 运行环境 */
  environment: RuntimeEnvironment
  projectName: string
  /** @kucoin-base/storage */
  kcStorage?: Storage
  kcSessionStorage?: Storage
  JsBridge?: any
  sentry?: any
  /** 指定哪些url需要采集设备指纹 */
  fingerPrintUrlList?: FingerPrintUrl[]
  /** 上报库 */
  reporter?: any
  /** 动态加载kc-report */
  getReporter?: () => Promise<any>
  csrfToken?: string
  /** 当前页面的语言 */
  langByPath?: string
  /** 直接指定x-version，不指定则从 storage 获取 */
  xVersion?: string
  /** 自动开启重试，只在ssr环境生效，默认true */
  autoRetry?: boolean
  // 其他配置
  // [key: string]: any;
}

export type Response<T = any> = {
  success: boolean
  code: string
  msg: string
  retry: boolean
} & T

export type AxiosInterceptor<T> = {
  /** 拦截器名称，可以按照name移除 */
  name: string
  description?: string
  /**
   * 拦截器优先级，数字越大越先执行，默认：1
   * 默认情况下，axios请求拦截器先注册的后执行，响应拦截器先注册的先执行
   * 这里提供统一的逻辑：优先级大的先执行，优先级相同的先注册的先执行
   * */
  priority?: number
  onFulfilled?: (value: T) => T | Promise<T>
  onRejected?: (error: any) => any
  /** 执行条件 */
  runWhen?: (config: T) => boolean
  /** 支持的运行时环境，不配置表示通用 */
  supportEnvironment?: RuntimeEnvironment[]
  /** 默认预置的运行时环境，不配置都预置 */
  presetEnvironment?: RuntimeEnvironment[]
}
// ...
export type RequestConfigWithExtra = InternalAxiosRequestConfig &
  RequestExtraConfig

export type RequestInterceptorObject = AxiosInterceptor<RequestConfigWithExtra>

export type RequestInterceptor =
  | RequestInterceptorObject
  | ((config: CreateHttpClientConfig) => RequestInterceptorObject)


export type ResponseType = AxiosResponse & { config: RequestConfigWithExtra }

export type ResponseInterceptorObject = AxiosInterceptor<ResponseType>

export type ResponseInterceptor =
  | ResponseInterceptorObject
  | ((config: CreateHttpClientConfig) => ResponseInterceptorObject)